Changed loader disk reference
Loader no longer uses a copy of the disk object; it is just passed in by reference when it needs to run.  This may not need to be a class actually, maybe it could just be a function.  We may want to split it into multiple functions so it's easier to understand.

Next TODO:  Create a PCB upon loading the program file.