Changed number of CPUs. Made CPUs global instead.
The cpus are now global.  They need to have variables that hold the current
process in order to keep track of the running one.  We don't want to use a
Queue for running processes because we don't know when a process will exit.

Separated out some OSDriver functions to make it easier to read.

Dispatcher now changes the status of a process that is running to RUNNING.  We
then pop the process from the ready_queue.  This lets us give different
processes to different CPUs.